Refactor bulk of audio-related logic into useToneAudio hook
